requirements.txt
absl-py==0.8.0
asn1crypto==0.24.0
astor==0.8.0
certifi==2019.9.11
cycler==0.10.0
gast==0.2.2
google-pasta==0.1.7
grpcio==1.23.0
h5py==2.9.0
idna==2.6
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.1.0
Markdown==3.1.1
matplotlib==3.1.1
numpy==1.17.1
opt-einsum==3.0.1
Pillow==6.1.0
pipenv==2018.11.26
protobuf==3.9.1
pycrypto==2.6.1
pygobject==3.26.1
pyparsing==2.4.2
python-apt==1.6.4
python-dateutil==2.8.0
pyxdg==0.25
PyYAML==5.1.2
SecretStorage==2.3.1
six==1.11.0
tb-nightly==1.15.0a20190806
tensorboard==2.0.0
tensorflow-estimator==2.0.0
tensorflow-gpu==2.0.0
termcolor==1.1.0
tf-estimator-nightly==1.14.0.dev2019080601
tqdm==4.36.1
virtualenv==16.7.5
virtualenv-clone==0.5.3
Werkzeug==0.15.5
wrapt==1.11.2
